requirements.txt
torch>=2.2.0
diffusers>=0.27.0
transformers>=4.40.0
accelerate>=0.29.0
safetensors>=0.4.0
Pillow>=10.0.0
